requirements.txt
cffi
simplekv
discord.py
steam
motor
tornado
thegamesdb
markovify
feedparser
wikipedia
